+++ eNMS/controller/base.py
@@ -334,9 +334,9 @@ class BaseController:
         }
 
     def compare(self, type, device_name, v1, v2):
-        if type == "result":
-            first = self.str_dict(getattr(db.fetch(type, id=v1), "result"))
-            second = self.str_dict(getattr(db.fetch(type, id=v2), "result"))
+        if type in ("result", "device_result"):
+            first = self.str_dict(getattr(db.fetch("result", id=v1), "result"))
+            second = self.str_dict(getattr(db.fetch("result", id=v2), "result"))
         else:
             first = self.get_git_network_data(device_name, v1)[type]
             second = self.get_git_network_data(device_name, v2)[type]